+++ com.unity.ml-agents/Runtime/Academy.cs
@@ -25,26 +25,6 @@ using Unity.Barracuda;
 
 namespace Unity.MLAgents
 {
-    internal class RecursionChecker : IDisposable
-    {
-        private bool m_IsRunning;
-
-        public IDisposable Start()
-        {
-            if (m_IsRunning)
-            {
-                throw new UnityAgentsException("Don't do this.");
-            }
-            m_IsRunning = true;
-            return this;
-        }
-
-        public void Dispose()
-        {
-            m_IsRunning = false;
-        }
-    }
-
     /// <summary>
     /// Helper class to step the Academy during FixedUpdate phase.
     /// </summary>
